mega.co.nz: initial support for account upload
This is made as a plowshare plugin because it requires a custom executable to work.
It depends on libcrypto (openssl) because MEGA is relying on crypto (AES, RSA).
Without this executable, plowshare mega module won't work at all.

So, this is a basic account upload and file is make public.

Random notes:
- Had to limit account password length to 16 characters. Need more work for greater passwords
- File is not shown on "File manager", but you can download it normally with firefox.